⬆️ Order Customer Update
Type(s): UPDATE

Issue(s):
- JIRA: https://blabol.atlassian.net/browse/MSBE-766

+++ app/Services/Order/OrderService.php
@@ -31,13 +31,6 @@ class OrderService extends BaseService
         $this->paymentLinkRepository = $paymentLinkRepository;
     }
 
-    public function checkOrder($partner_id, $order_id)
-    {
-        $order = $this->orderRepositoryInterface->where('partner_id', $partner_id)->find($order_id);
-        if(!$order) return $this->error('অর্ডারটি পাওয়া যায় নি', 404);
-        return $order;
-    }
-
     public function getOrderList($partner_id, $request)
     {
         list($offset, $limit) = calculatePagination($request);
@@ -52,13 +45,14 @@ class OrderService extends BaseService
 
         $ordersList = $this->orderRepositoryInterface->getOrderListWithPagination($offset, $limit, $partner_id, $orderSearch, $orderFilter);
         $orderList = OrderResource::collection($ordersList);
-        if(!$orderList) return $this->error('অর্ডারটি পাওয়া যায় নি', 404);
+        if(!$orderList) return $this->error("You're not authorized to access this order", 403);
         else return $this->success('Success', ['orders' => $orderList], 200, true);
     }
 
     public function getOrderDetails($partner_id, $order_id)
     {
-        $order = $this->checkOrder($partner_id, $order_id);
+        $order = $this->orderRepositoryInterface->where('partner_id', $partner_id)->find($order_id);
+        if(!$order) return $this->error("You're not authorized to access this order", 403);
         $order->calculate();
         $resource = new OrderWithProductResource($order);
         if($order->due > 0){
@@ -71,7 +65,8 @@ class OrderService extends BaseService
 
     public function update($orderUpdateRequest, $partner_id, $order_id)
     {
-        $orderDetails = $this->checkOrder($partner_id, $order_id);
+        $orderDetails = $this->orderRepositoryInterface->where('partner_id', $partner_id)->find($order_id);
+        if(!$orderDetails) return $this->error("You're not authorized to access this order", 403);
         $this->updater->setPartnerId($partner_id)
             ->setOrderId($order_id)
             ->setOrder($orderDetails)
@@ -94,7 +89,8 @@ class OrderService extends BaseService
 
     public function delete($partner_id, $order_id)
     {
-        $order = $this->checkOrder($partner_id, $order_id);
+        $order = $this->orderRepositoryInterface->where('partner_id', $partner_id)->find($order_id);
+        if(!$order) return $this->error("You're not authorized to access this order", 403);
         $OrderSkusIds = $this->orderSkusRepositoryInterface->where('order_id', $order_id)->get(['id']);
         $this->orderSkusRepositoryInterface->whereIn('id', $OrderSkusIds)->delete();
         $order->delete();
@@ -104,7 +100,7 @@ class OrderService extends BaseService
     public function getOrderWithChannel($order_id)
     {
         $orderDetails = $this->orderRepositoryInterface->find($order_id);
-        if(!$orderDetails) return $this->error('অর্ডারটি পাওয়া যায় নি', 404);
+        if(!$orderDetails) return $this->error("You're not authorized to access this order", 403);
         $order = [
             'id' => $orderDetails->id,
             'sales_channel' => $orderDetails->sales_channel_id == 1 ? 'pos' : 'webstore'
@@ -115,9 +111,6 @@ class OrderService extends BaseService
     public function getOrderPaymentLink(Order $order) {
         $payment_link_target = $order->getPaymentLinkTarget();
         $payment_link = $this->paymentLinkRepository->getActivePaymentLinkByPosOrder($payment_link_target);
-        if ($payment_link) {
-            return $payment_link;
-        } else
-            return false;
+        return isset($payment_link) ? $payment_link : false;
     }
 }
